requirements.txt
Flask
python-dotenv
pymongo
Flask-Session
redis